+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 1.2.1
+ - `context.showLoaderOverlay` and `context.hideLoaderOverlay` marked as depricated. You should use `context.loaderOverlay.show` and `context.loaderOverlay.hide`
+ - widget that you pass to `context.loaderOverlay.show` will be shown instead of `widgetOverlay`
+ - Added `context.loaderOverlay.visible`
+ - Added `context.loaderOverlay.overlayWidgetType`
+ - `useDefaultLoading` set to `true` by default
+
 # 1.2.0
 
 - Adding possibility to pass a widget to show underneath the loading indicator